Treat component store the same as other stores, and move component utils into helpers
