+++ packages/builder/src/pages/builder/app/[application]/design/[screenId]/[componentId]/_components/Component/ComponentSettingsPanel.svelte
@@ -9,7 +9,7 @@
   import DesignSection from "./DesignSection.svelte"
   import CustomStylesSection from "./CustomStylesSection.svelte"
   import ConditionalUISection from "./ConditionalUISection.svelte"
-  import { getComponentName } from "stores/builder/components/utils"
+  import { getComponentName } from "helpers/components"
   import {
     getBindableProperties,
     getComponentBindableProperties,
