+++ packages/builder/src/pages/builder/app/[application]/design/[screenId]/_components/ComponentList/ComponentTree.svelte
@@ -14,7 +14,7 @@
     findComponentPath,
     getComponentText,
     getComponentName,
-  } from "stores/builder/components/utils"
+  } from "helpers/components"
   import { get } from "svelte/store"
   import { dndStore } from "./dndStore"
